create spinner component
